fixing validation errors in registration
