+++ ClinicWeb/src/main/java/pl/infoshare/clinicweb/doctor/Doctor.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 import pl.infoshare.clinicweb.patient.Address;
 import pl.infoshare.clinicweb.patient.Patient;
 import pl.infoshare.clinicweb.patientCard.PatientCard;
-import pl.infoshare.clinicweb.user.PersonDetails;
+import pl.infoshare.clinicweb.user.entity.PersonDetails;
 
 import java.util.List;
