+++ ClinicWeb/src/main/java/pl/infoshare/clinicweb/doctor/DoctorMapper.java
@@ -1,7 +1,7 @@
 package pl.infoshare.clinicweb.doctor;
 
 import org.springframework.stereotype.Component;
-import pl.infoshare.clinicweb.user.PersonDetails;
+import pl.infoshare.clinicweb.user.entity.PersonDetails;
 
 
 @Component
