+++ ClinicWeb/src/main/java/pl/infoshare/clinicweb/patient/PatientController.java
@@ -12,7 +12,7 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import pl.infoshare.clinicweb.advice.ExceptionHandlerApp;
 import pl.infoshare.clinicweb.doctor.DoctorDto;
 import pl.infoshare.clinicweb.doctor.DoctorService;
-import pl.infoshare.clinicweb.user.PersonDetails;
+import pl.infoshare.clinicweb.user.entity.PersonDetails;
 import pl.infoshare.clinicweb.advice.PeselFormatException;
 import pl.infoshare.clinicweb.user.Utils;
