+++ ClinicWeb/src/main/java/pl/infoshare/clinicweb/patient/PatientMapper.java
@@ -1,7 +1,7 @@
 package pl.infoshare.clinicweb.patient;
 
 import org.springframework.stereotype.Component;
-import pl.infoshare.clinicweb.user.PersonDetails;
+import pl.infoshare.clinicweb.user.entity.PersonDetails;
 
 @Component
 public class PatientMapper {
