+++ ClinicWeb/src/main/java/pl/infoshare/clinicweb/user/entity/Gender.java
@@ -1,4 +1,4 @@
-package pl.infoshare.clinicweb.user;
+package pl.infoshare.clinicweb.user.entity;
 
 import com.fasterxml.jackson.annotation.JsonValue;
