+++ ClinicWeb/src/main/java/pl/infoshare/clinicweb/user/login/LoginController.java
@@ -1,0 +1,46 @@
+package pl.infoshare.clinicweb.user.login;
+
+import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import pl.infoshare.clinicweb.user.entity.AppUser;
+
+@Controller
+@AllArgsConstructor
+@Slf4j
+public class LoginController {
+
+    @GetMapping("/login")
+    public String login() {
+
+        AppUser user = getPrincipal();
+
+        if (user != null) {
+            log.info("User with email: {} was successfully logged in.", user.getEmail());
+            return "redirect:/index";
+
+        }
+
+        return "login";
+    }
+
+    @GetMapping("/logout")
+    public String logout() {
+
+        log.info("User was successfully logged out.");
+        return "index";
+
+    }
+
+    private AppUser getPrincipal() {
+        AppUser user = null;
+        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal() instanceof AppUser) {
+            user = (AppUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        }
+        return user;
+    }
+
+
+}
