+++ ClinicWeb/src/main/java/pl/infoshare/clinicweb/user/registration/AppUserDto.java
@@ -1,11 +1,13 @@
-package pl.infoshare.clinicweb.user;
+package pl.infoshare.clinicweb.user.registration;
 
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Size;
 import lombok.Data;
 import pl.infoshare.clinicweb.passwordAnnotation.PasswordMatcherValidator;
+import pl.infoshare.clinicweb.user.entity.Role;
 
-import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.Pattern;
-import javax.validation.constraints.Size;
 
 @Data
 @PasswordMatcherValidator
@@ -13,7 +15,7 @@ public class AppUserDto {
 
     private Long id;
     @NotEmpty(message = "Pole email nie może być puste.")
-    @Pattern(regexp = ("/^\\S+@\\S+\\.\\S+$/"))
+    @Pattern(regexp = ("/^\\S+@\\S+\\.\\S+$/"), message = "Niepoprawny format, poprawny format e-mail np. xxxx@xxx.xx")
     private String email;
     @NotEmpty(message = "Pole hasło nie może być puste.")
     @Size(min = 6, message = "Hasło musi składać się z przynajmniej 6 znaków.")
@@ -21,6 +23,6 @@ public class AppUserDto {
     @NotEmpty(message = "Pole nie może być puste.")
     @Size(min = 6, message = "Hasło musi składać się z przynajmniej 6 znaków.")
     private String confirmPassword;
-    @NotEmpty(message = "Podaj swoją rolę użytkownika:")
+    @NotNull(message = "Podaj swoją rolę użytkownika:")
     private Role role;
 }
