+++ ClinicWeb/src/main/java/pl/infoshare/clinicweb/user/registration/RegistrationController.java
@@ -1,23 +1,22 @@
-package pl.infoshare.clinicweb.user;
+package pl.infoshare.clinicweb.user.registration;
 
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
+import pl.infoshare.clinicweb.user.service.AppUserService;
 
 @Controller
 @AllArgsConstructor
 @Slf4j
-public class LoginController {
+public class RegistrationController {
 
     private final AppUserService userService;
-
 
     @GetMapping("/register")
     public String registerForm(Model model) {
@@ -29,11 +28,11 @@ public class LoginController {
     }
 
     @PostMapping("/register")
-    public String registerFormSubmission(@Valid AppUserDto user, BindingResult bindingResult, Model model) {
-
-        model.addAttribute("user", user);
+    public String registerFormSubmission(@Valid @ModelAttribute("user") AppUserDto user, BindingResult bindingResult, Model model) {
 
         if (bindingResult.hasErrors()) {
+
+            model.addAttribute("user", user);
             log.info("Validation error occured when registering user.");
             return "user/registry";
         }
@@ -44,36 +43,4 @@ public class LoginController {
 
         return "redirect:/registry";
     }
-
-    @GetMapping("/login")
-    public String login() {
-
-        AppUser user = getPrincipal();
-
-        if (user != null) {
-            log.info("User with id: {} was successfully logged in.", user.getId());
-            return "redirect:/index";
-
-        }
-
-        return "login";
-    }
-
-    @GetMapping("/logout")
-    public String logout() {
-
-        log.info("User was successfully logged out.");
-        return "index";
-
-    }
-
-    private AppUser getPrincipal() {
-        AppUser user = null;
-        if (SecurityContextHolder.getContext().getAuthentication().getPrincipal() instanceof AppUser) {
-            user = (AppUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        }
-        return user;
-    }
-
-
 }
