+++ ClinicWeb/src/main/java/pl/infoshare/clinicweb/visit/VisitMapper.java
@@ -4,17 +4,13 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Component;
 import pl.infoshare.clinicweb.doctor.Doctor;
 import pl.infoshare.clinicweb.patient.Patient;
-import pl.infoshare.clinicweb.user.PersonDetails;
-
+import pl.infoshare.clinicweb.user.entity.PersonDetails;
 
 
 @Component
 @AllArgsConstructor
 
 public class VisitMapper {
-
-
-
 
     public VisitDto toVisitDto(Visit visit) {
 
@@ -52,7 +48,6 @@ public class VisitMapper {
         if (visitDto == null) {
             throw new IllegalArgumentException("VisitDto cannot be null");
         }
-
 
 
         Patient patient = new Patient();
