+++ ClinicWeb/src/main/resources/templates/user/registry.html
@@ -121,15 +121,13 @@
 
             <div class="input-group">
                 <input type="password" name="password" th:field="*{password}" placeholder="Hasło" required>
-                <p th:if="${#fields.hasErrors('password')}" class="text-danger"
-                   th:errors="*{password}">Password error</p>
             </div>
 
             <div class="input-group">
                 <input type="password" th:field="*{confirmPassword}" name="confirmPassword" placeholder="Powtórz hasło"
                        required>
-                <p th:if="${#fields.hasErrors('confirmPassword')}" class="text-danger"
-                   th:errors="*{confirmPassword}">Confirm password error</p>
+                <p th:if="${#fields.hasErrors('password')}" class="text-danger"
+                   th:errors="${user.password}">Confirm password error</p>
             </div>
 
             <select name="role"
@@ -138,13 +136,13 @@
                     oninvalid="setCustomValidity('Rejestruje się jako')"
                     oninput="setCustomValidity('')">
                 <option value="" selected>--Wybierz rolę--</option>
-                <option th:each="role : ${T(pl.infoshare.clinicweb.user.Role).values()}"
+                <option th:each="role : ${T(pl.infoshare.clinicweb.user.entity.Role).values()}"
                         th:if="${role.name} != 'ADMIN'"
                         th:value="${role}"
                         th:text="${role.getRoleDescription()}"></option>
             </select>
             <p th:if="${#fields.hasErrors('role')}" class="text-danger"
-               th:errors="*{role}">Role error</p>
+               th:errors="${user.role}">Role error</p>
 
             <input type="submit" value="Zarejestruj się">
         </div>
